Revert "[backend] refactor #35"

+++ backend/internal/service/skill_service.go
@@ -3,29 +3,33 @@ package service
 import (
 	"HistoryHub/internal/model"
 	"HistoryHub/internal/repository"
+	"errors"
 	"time"
-
-	"github.com/google/uuid"
 )
 
-func GetSkillsByUserID(UserID uuid.UUID) ([]model.Skill, error) {
-	skills, err := repository.GetSkillsByUserID(UserID)
+func CreateSkill(UserID string, skill model.Skill) error {
+	profile, err := repository.GetProfileByUserID(UserID)
 	if err != nil {
-		return nil, err
+		return err
 	}
-	return skills, nil
-}
 
-func CreateSkill(skill model.Skill) error {
+	skill.ProfileID = profile.ID
 	skill.CreatedAt = time.Now()
 	skill.UpdatedAt = time.Now()
+
 	if err := repository.CreateSkill(skill); err != nil {
 		return err
 	}
 	return nil
 }
 
-func UpdateSkill(skill model.Skill) error {
+func UpdateSkill(UserID string, skill model.Skill) error {
+	profile, _ := repository.GetProfileByUserID(UserID)
+	preSkill, _ := repository.GetSkillByID(skill.ID)
+	if profile.ID != preSkill.ProfileID {
+		return errors.New("skill not yours")
+	}
+	skill.ProfileID = profile.ID
 	skill.UpdatedAt = time.Now()
 	if err := repository.UpdateSkill(skill); err != nil {
 		return err
@@ -33,9 +37,26 @@ func UpdateSkill(skill model.Skill) error {
 	return nil
 }
 
-func DeleteSkill(id uuid.UUID) error {
+func DeleteSkill(UserID string, id uint) error {
+	profile, _ := repository.GetProfileByUserID(UserID)
+	preSkill, _ := repository.GetSkillByID(id)
+	if profile.ID != preSkill.ProfileID {
+		return errors.New("skill not yours")
+	}
 	if err := repository.DeleteSkillByID(id); err != nil {
 		return err
 	}
 	return nil
+}
+
+func GetSkills(UserID string) ([]model.Skill, error) {
+	profile, err := repository.GetProfileByUserID(UserID)
+	if err != nil  {
+		return nil, err
+	}
+	skills, err := repository.GetSkillsByID(profile.ID)
+	if err != nil  {
+		return nil, err
+	}
+	return skills, nil
 }
